Make openstack-congress documentation conditional
It lets downstream the possibility to skip building docs.

I tried to fix de FIXME by adding -W to sphinx-build command
but we still can not add it.
A FFE [1] has been raised and granted following the bug [2].
They bumped sphinx to 1.7.3 in this change [3].
But a couple of weeks ago they took a step back [4].
Should we open the bug again ?

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-August/133065.html
[2] https://bugs.launchpad.net/congress/+bug/1778171
[3] https://review.opendev.org/#/c/589995/
[4] https://review.opendev.org/#/c/666013/

Change-Id: I0ad99fda1f0510784a501debb98cccd0c856f76c